requirements.txt
requests
fastapi
uvicorn
starlette-prometheus
Authlib
httpx
itsdangerous
python-jose[cryptography]
